Task 1 - Created Revenue Metric Card
Use getElementById to pull the document from the html file by its ID. Use the querySelector to pull the document by its first query instance in the html. Create a new element revMetricCard that is stored to the document in a div container. Set the attributes of class as metric card and id as revenue card for the element. Populate the element with the title of revenue and the starting balance of $0. Add the element to the dashboard.