Task 2 - Updated Metric Cards via Array Conversion
Create the additional metric cards for profits and expenses using the same steps as task 1.  Create a list of all metric cards by using the query selector and assigning them to the nodelist element. Convert the list to an array and for each array, add the text "updated" to the content printed in the html file and change the color of text printed to dark green.