Task 3 - Implemented Dynamic Inventory List
Create a button that when clicked, allows us to add a new product to the inventory list. The inventory list and the product button are displayed to the html by pulling the commands through their ids. In the function, create another function that removed the new product from the inventory list when it is clicked on. Link the list and the new product together.