Task 4 - Demonstrated Event Bubbling in Customer Section
Pull the parent container by the ID and the customer card by ID from the html. For when the customer section is clicked, log to the console that it was clicked. For each item in the customer cards, when an item is clicked, log a message saying that the customer card is clicked. If the stop propagation even is active, only the customer card is clicked. If it is not active, it will click both the customer card and the customer section.